Externally bundled natives and hopefully partial test setup
