+++ .github/workflows/release.yml
@@ -64,19 +64,6 @@
                     "run": "git push && git push --tags"
                 },
                 {
-                    "with": {
-                        "gradle-build-module": ":\n:opensesame-compile\n:opensesame-core\n:opensesame-fabric\n:opensesame-groovy\n:opensesame-javac\n:opensesame-mixin\n:opensesame-plugin-core\n:opensesame-plugin-loom\n:opensesame-core:natives",
-                        "gradle-build-configuration": "compileClasspath",
-                        "sub-module-mode": "INDIVIDUAL_DEEP",
-                        "include-build-environment": true
-                    },
-                    "name": "Submit Dependencies",
-                    "uses": "mikepenz/gradle-dependency-submission@v0.9.0",
-                    "env": {
-                        "GITHUB_TOKEN": "${{ secrets.GITHUB_TOKEN }}"
-                    }
-                },
-                {
                     "name": "Record Version",
                     "run": "./gradlew recordVersion",
                     "id": "record_version",
@@ -88,6 +75,13 @@
                     "name": "Capture Recorded Version",
                     "run": "echo version=$(cat build/recordVersion.txt) >> \"$GITHUB_OUTPUT\"",
                     "id": "record_version_capture_version"
+                },
+                {
+                    "name": "Submit Dependencies",
+                    "uses": "gradle/actions/dependency-submission@v3",
+                    "env": {
+                        "GITHUB_TOKEN": "${{ secrets.GITHUB_TOKEN }}"
+                    }
                 }
             ]
         },
