+++ .github/workflows/report.yml
@@ -43,9 +43,22 @@
                         "path": "**/test-results/testOn21/**/TEST-*.xml",
                         "reporter": "java-junit",
                         "fail-on-empty": "true",
-                        "list-tests": "failed"
+                        "list-tests": "failed",
+                        "list-suites": "failed"
                     },
                     "name": "JUnit Test Report - Java 21",
+                    "uses": "dorny/test-reporter@v1"
+                },
+                {
+                    "with": {
+                        "name": "Test Results - Java 22",
+                        "path": "**/test-results/testOn22/**/TEST-*.xml",
+                        "reporter": "java-junit",
+                        "fail-on-empty": "true",
+                        "list-tests": "failed",
+                        "list-suites": "failed"
+                    },
+                    "name": "JUnit Test Report - Java 22",
                     "uses": "dorny/test-reporter@v1"
                 },
                 {
