--- core/src/main/java/dev/lukebemish/opensesame/runtime/NativeImplementations.java
@@ -1,7 +1,0 @@
-package dev.lukebemish.opensesame.runtime;
-
-import java.lang.invoke.MethodHandles;
-
-class NativeImplementations {
-    native static MethodHandles.Lookup nativeImplLookup();
-}
