+++ natives/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module dev.lukebemish.opensesame.natives {
+    requires static org.jetbrains.annotations;
+
+    opens dev.lukebemish.opensesame.natives to dev.lukebemish.opensesame.core;
+    exports dev.lukebemish.opensesame.natives to dev.lukebemish.opensesame.core;
+}
