+++ settings.gradle
@@ -15,7 +15,7 @@ rootProject.name = 'opensesame'
 
 def subprojects = [
         'core',
-        'core:natives',
+        'natives',
         'compile',
         'groovy',
         'javac',
